requirements.txt
alembic==1.7.6
cffi==1.15.0
click==8.0.4
colorama==0.4.4
cryptography==36.0.1
Flask==2.0.3
Flask-CKEditor==0.4.6
Flask-Login==0.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
mysql-connector==2.2.9
mysql-connector-python==8.0.28
mysql-connector-python-rf==2.2.2
protobuf==3.19.4
psycopg2==2.9.3
pycparser==2.21
PyMySQL==1.0.2
SQLAlchemy==1.4.31
Werkzeug==2.0.3
WTForms==3.0.1
